Move all basic markdown rules into the markdown common package
