+++ markymark-android/src/main/java/com/m2mobi/markymarkandroid/ParagraphDisplayItem.kt
@@ -12,7 +12,7 @@ import android.widget.TextView
 
 import com.m2mobi.markymark.DisplayItem
 import com.m2mobi.markymark.InlineConverter
-import com.m2mobi.markymarkcommon.markdownitems.Paragraph
+import com.m2mobi.markymarkcommon.markdownitem.Paragraph
 
 /**
  * Converts an [Paragraph] to a [View].
