+++ markymark-android/src/main/java/com/m2mobi/markymarkandroid/inline/ItalicInlineDisplayItem.kt
@@ -10,7 +10,7 @@ import android.text.style.StyleSpan
 
 import com.m2mobi.markymark.InlineConverter
 import com.m2mobi.markymark.InlineDisplayItem
-import com.m2mobi.markymarkcommon.markdownitems.inline.ItalicString
+import com.m2mobi.markymarkcommon.markdownitem.inline.ItalicString
 
 /**
  * Converts an [ItalicString] to a [Spanned].
