+++ markymark-common/src/main/java/com/m2mobi/markymarkcommon/markdownitem/inline/InlineString.java
@@ -2,7 +2,7 @@
 * Copyright (C) M2mobi BV - All Rights Reserved
 */
 
-package com.m2mobi.markymarkcommon.markdownitems.inline;
+package com.m2mobi.markymarkcommon.markdownitem.inline;
 
 import com.m2mobi.markymark.item.inline.MarkdownString;
