+++ markymark-common/src/main/java/com/m2mobi/markymarkcommon/rule/HeaderRule.java
@@ -22,42 +22,39 @@
  * SOFTWARE.
  */
 
-package com.m2mobi.markymarkcontentful.rules;
+package com.m2mobi.markymarkcommon.rule;
 
-import com.m2mobi.markymarkcommon.rule.ImageRule;
+import com.m2mobi.markymark.item.MarkdownItem;
+import com.m2mobi.markymark.rules.RegexRule;
+import com.m2mobi.markymark.rules.Rule;
+import com.m2mobi.markymarkcommon.markdownitem.Header;
 
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-
-import java.util.ArrayList;
 import java.util.List;
-
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 /**
- * Tests for {@link ImageRule}
+ * {@link Rule} that matches headers from H1 to H6
  */
-public class ImageRuleTest {
+public class HeaderRule extends RegexRule {
 
-	private ImageRule mImageRule = null;
+	/** Pattern used to find headers and the header size */
+	private static final Pattern HEADER_PATTERN = Pattern.compile("^(#{1,6})\\s*(.+)$");
 
-	@BeforeEach
-	public void init() {
-		mImageRule = new ImageRule();
+	@Override
+	protected Pattern getRegex() {
+		return HEADER_PATTERN;
 	}
 
-	@Test
-	public void shouldBeImage() {
-		List<String> strings = new ArrayList<>();
-		strings.add("![Image](www.google.com/images/cheese)");
-		assertTrue(mImageRule.conforms(strings));
-	}
-
-	@Test
-	public void shouldNotBeImage() {
-		List<String> strings = new ArrayList<>();
-		strings.add("[Image](www.google.com/images/cheese)");
-		assertFalse(mImageRule.conforms(strings));
+	@Override
+	public MarkdownItem toMarkdownItem(final List<String> pMarkdownLines) {
+		String headerContent = "";
+		int headerType = 1;
+		Matcher headerMatcher = HEADER_PATTERN.matcher(pMarkdownLines.get(0));
+		if (headerMatcher.matches()) {
+			headerType = headerMatcher.group(1).length();
+			headerContent = headerMatcher.group(2);
+		}
+		return new Header(headerContent, headerType);
 	}
 }
