+++ markymark-contentful/src/main/java/com/m2mobi/markymarkcontentful/ContentfulFlavor.java
@@ -7,18 +7,18 @@ package com.m2mobi.markymarkcontentful;
 import com.m2mobi.markymark.Flavor;
 import com.m2mobi.markymark.rules.InlineRule;
 import com.m2mobi.markymark.rules.Rule;
-import com.m2mobi.markymarkcontentful.rules.CodeBlockRule;
-import com.m2mobi.markymarkcontentful.rules.HeaderRule;
-import com.m2mobi.markymarkcontentful.rules.HorizontalLineRule;
-import com.m2mobi.markymarkcontentful.rules.ImageRule;
-import com.m2mobi.markymarkcontentful.rules.ListRule;
-import com.m2mobi.markymarkcontentful.rules.ParagraphRule;
-import com.m2mobi.markymarkcontentful.rules.QuoteRule;
-import com.m2mobi.markymarkcontentful.rules.inline.BoldRule;
-import com.m2mobi.markymarkcontentful.rules.inline.InlineCodeRule;
-import com.m2mobi.markymarkcontentful.rules.inline.ItalicRule;
-import com.m2mobi.markymarkcontentful.rules.inline.LinkRule;
-import com.m2mobi.markymarkcontentful.rules.inline.StrikeRule;
+import com.m2mobi.markymarkcommon.rule.CodeBlockRule;
+import com.m2mobi.markymarkcommon.rule.HeaderRule;
+import com.m2mobi.markymarkcommon.rule.HorizontalLineRule;
+import com.m2mobi.markymarkcommon.rule.ImageRule;
+import com.m2mobi.markymarkcommon.rule.ListRule;
+import com.m2mobi.markymarkcommon.rule.ParagraphRule;
+import com.m2mobi.markymarkcommon.rule.QuoteRule;
+import com.m2mobi.markymarkcommon.rule.inline.BoldRule;
+import com.m2mobi.markymarkcommon.rule.inline.InlineCodeRule;
+import com.m2mobi.markymarkcommon.rule.inline.ItalicRule;
+import com.m2mobi.markymarkcommon.rule.inline.LinkRule;
+import com.m2mobi.markymarkcommon.rule.inline.StrikeRule;
 
 import java.util.ArrayList;
 import java.util.List;
