+++ markymark-contentful/src/test/java/com/m2mobi/markymarkcontentful/rules/HorizontalLineRuleTest.java
@@ -24,6 +24,8 @@
 
 package com.m2mobi.markymarkcontentful.rules;
 
+import com.m2mobi.markymarkcommon.rule.HorizontalLineRule;
+
 import org.junit.jupiter.api.Test;
 
 import java.util.ArrayList;
