+++ markymark-contentful/src/test/java/com/m2mobi/markymarkcontentful/rules/QuoteRuleTest.java
@@ -24,6 +24,8 @@
 
 package com.m2mobi.markymarkcontentful.rules;
 
+import com.m2mobi.markymarkcommon.rule.QuoteRule;
+
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
