+++ markymark-contentful/src/test/java/com/m2mobi/markymarkcontentful/rules/inline/ItalicRuleTest.java
@@ -24,6 +24,8 @@
 
 package com.m2mobi.markymarkcontentful.rules.inline;
 
+import com.m2mobi.markymarkcommon.rule.inline.ItalicRule;
+
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
