+++ markymark-contentful/src/test/java/com/m2mobi/markymarkcontentful/rules/inline/LinkRuleTest.java
@@ -24,7 +24,8 @@
 
 package com.m2mobi.markymarkcontentful.rules.inline;
 
-import com.m2mobi.markymarkcommon.markdownitems.inline.LinkString;
+import com.m2mobi.markymarkcommon.markdownitem.inline.LinkString;
+import com.m2mobi.markymarkcommon.rule.inline.LinkRule;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
