+++ markymark-contentful/src/test/java/com/m2mobi/markymarkcontentful/rules/inline/StrikeRuleTest.java
@@ -24,6 +24,8 @@
 
 package com.m2mobi.markymarkcontentful.rules.inline;
 
+import com.m2mobi.markymarkcommon.rule.inline.StrikeRule;
+
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
